+++ lib/speed_up_flutter.dart
@@ -3,3 +3,4 @@ library speed_up_flutter;
 export './src/extensions.dart';
 export './src/funcs.dart';
 export './src/section.dart';
+export './src/widgets/widgets.dart';